Move TokenImg to it's common components
