+++ src/hooks/useBetterAddTokenModal.tsx
@@ -3,7 +3,7 @@ import Modali, { useModali, ModalHook } from 'modali'
 import { fetchTokenData, FetchTokenResult } from 'services'
 import { Deferred, createDeferredPromise } from 'utils'
 import { TokenDetails } from 'types'
-import TokenImg from '../components/TokenImg'
+import TokenImg from '../components/common/TokenImg'
 import styled from 'styled-components'
 import { tokenListApi } from 'api'
 import { TokenFromExchange } from 'services/factories'
